refactor: inline `@sanity/insert-menu` into `@sanity/visual-editing`
Vendor the InsertMenu component into `@sanity/visual-editing` and define
`InsertMenuOptions` locally in `@sanity/visual-editing-types`, so neither
package depends on the standalone `@sanity/insert-menu` anymore.

The `@sanity/insert-menu` workspace is retired (reduced to a deprecation
README) and is no longer published, following the `@sanity/presentation`
precedent. Its only remaining consumer, `sanity`, is being inlined
separately in sanity-io/sanity.
