+++ .changeset/inline-insert-menu.md
@@ -1,0 +1,5 @@
+---
+"@sanity/visual-editing": patch
+---
+
+Inline `@sanity/insert-menu` instead of depending on the standalone package
