+++ .changeset/insert-menu-options-local.md
@@ -1,0 +1,5 @@
+---
+"@sanity/visual-editing-types": patch
+---
+
+Define `InsertMenuOptions` locally instead of re-exporting it from `@sanity/insert-menu`
