--- packages/insert-menu/.gitignore
@@ -1,9 +1,0 @@
-*.local
-*.log
-*.tgz
-
-.DS_Store
-.workshop
-dist
-etc
-node_modules
