+++ packages/insert-menu/README.md
@@ -1,1 +1,6 @@
 # `@sanity/insert-menu`
+
+> [!WARNING]
+> This package is deprecated and is no longer published.
+
+The `InsertMenu` component has been inlined into its only consumers, [`@sanity/visual-editing`](https://github.com/sanity-io/visual-editing/tree/main/packages/visual-editing) and [`sanity`](https://github.com/sanity-io/sanity). There is no standalone replacement.
