--- packages/insert-menu/package.json
@@ -1,78 +1,0 @@
-{
-  "name": "@sanity/insert-menu",
-  "version": "3.0.8",
-  "description": "",
-  "keywords": [],
-  "homepage": "https://github.com/sanity-io/visual-editing/tree/main/packages/insert-menu#readme",
-  "bugs": {
-    "url": "https://github.com/sanity-io/visual-editing/issues"
-  },
-  "license": "MIT",
-  "author": "Sanity.io <hello@sanity.io>",
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/sanity-io/visual-editing.git",
-    "directory": "packages/insert-menu"
-  },
-  "files": [
-    "dist"
-  ],
-  "type": "module",
-  "sideEffects": false,
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "source": "./src/index.ts",
-      "default": "./dist/index.js"
-    },
-    "./package.json": "./package.json"
-  },
-  "publishConfig": {
-    "exports": {
-      ".": "./dist/index.js",
-      "./package.json": "./package.json"
-    }
-  },
-  "scripts": {
-    "build": "pkg build --strict --check --clean",
-    "dev": "pkg build --strict",
-    "dev:workshop": "workshop dev",
-    "lint": "eslint .",
-    "prepack": "turbo run build",
-    "ts:check": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@sanity/icons": "^3.7.4",
-    "@sanity/ui": "^3.2.0",
-    "lodash-es": "^4.18.1",
-    "react-is": "^19.2.7"
-  },
-  "devDependencies": {
-    "@sanity/pkg-utils": "catalog:",
-    "@sanity/tsconfig": "catalog:",
-    "@sanity/types": "catalog:",
-    "@sanity/ui-workshop": "^3.4.0",
-    "@types/lodash-es": "^4.17.12",
-    "@types/react": "catalog:",
-    "@types/react-is": "^19.2.0",
-    "@typescript-eslint/eslint-plugin": "catalog:",
-    "@typescript-eslint/parser": "catalog:",
-    "babel-plugin-react-compiler": "catalog:",
-    "eslint": "^8.57.1",
-    "eslint-plugin-import": "^2.32.0",
-    "lint-staged": "^15.5.2",
-    "react": "catalog:",
-    "react-dom": "catalog:",
-    "typescript": "catalog:"
-  },
-  "peerDependencies": {
-    "@sanity/types": "*",
-    "react": "^19.2"
-  },
-  "browserslist": "extends @sanity/browserslist-config",
-  "engines": {
-    "node": ">=20.19 <22 || >=22.12"
-  }
-}
