--- packages/insert-menu/src/__workshop__/.eslintrc
@@ -1,6 +1,0 @@
-{
-  "rules": {
-    "@typescript-eslint/explicit-module-boundary-types": "off",
-    "@typescript-eslint/no-unused-vars": "off",
-  },
-}
