--- packages/insert-menu/src/__workshop__/index.ts
@@ -1,14 +1,0 @@
-import {defineScope} from '@sanity/ui-workshop'
-import {lazy} from 'react'
-
-export default defineScope({
-  name: 'insert-menu',
-  title: 'Insert Menu',
-  stories: [
-    {
-      name: 'InsertMenu',
-      title: 'InsertMenu',
-      component: lazy(() => import('./full')),
-    },
-  ],
-})
