--- packages/insert-menu/src/index.ts
@@ -1,2 +1,0 @@
-export * from './InsertMenu'
-export * from './InsertMenuOptions'
