--- packages/insert-menu/tsconfig.dist.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.settings",
-  "compilerOptions": {
-    "rootDir": "src"
-  },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
-  "exclude": ["dist", "node_modules"]
-}
