--- packages/insert-menu/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "./tsconfig.settings",
-  "include": ["**/*.ts", "**/*.tsx"],
-  "exclude": ["dist", "./node_modules"]
-}
