--- packages/insert-menu/tsconfig.settings.json
@@ -1,7 +1,0 @@
-{
-  "extends": "@sanity/tsconfig/strictest",
-  "compilerOptions": {
-    "rootDir": ".",
-    "outDir": "./dist"
-  }
-}
