--- packages/insert-menu/turbo.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://turbo.build/schema.json",
-  "extends": ["//"],
-  "tasks": {
-    "build": {
-      "dependsOn": ["^build"]
-    },
-    "dev": {
-      "persistent": false
-    }
-  }
-}
