--- packages/insert-menu/workshop.config.ts
@@ -1,7 +1,0 @@
-import {defineConfig} from '@sanity/ui-workshop'
-import {buildTheme} from '@sanity/ui/theme'
-
-export default defineConfig({
-  theme: buildTheme(),
-  title: '@sanity/insert-menu',
-})
