+++ packages/visual-editing-types/package.json
@@ -46,7 +46,6 @@
     "@repo/eslint-config": "workspace:*",
     "@repo/package.config": "workspace:*",
     "@sanity/client": "catalog:",
-    "@sanity/insert-menu": "workspace:^",
     "@sanity/pkg-utils": "catalog:",
     "@sanity/tsconfig": "catalog:",
     "@sanity/types": "catalog:",
