+++ packages/visual-editing/package.json
@@ -103,14 +103,16 @@
   "dependencies": {
     "@sanity/comlink": "^4.0.1",
     "@sanity/icons": "^3.7.4",
-    "@sanity/insert-menu": "workspace:^",
     "@sanity/mutate": "^0.18.0",
     "@sanity/presentation-comlink": "^2.1.0",
     "@sanity/preview-url-secret": "workspace:^",
+    "@sanity/types": "catalog:",
     "@sanity/ui": "^3.2.0",
     "@sanity/visual-editing-csm": "workspace:^",
     "@vercel/stega": "^1.1.0",
     "dequal": "^2.0.3",
+    "lodash-es": "^4.18.1",
+    "react-is": "^19.2.7",
     "rxjs": "^7.8.2",
     "scroll-into-view-if-needed": "^3.1.0",
     "xstate": "^5.32.1"
@@ -135,8 +137,10 @@
     "@sveltejs/vite-plugin-svelte": "^6.2.4",
     "@tailwindcss/postcss": "catalog:tailwind3",
     "@tailwindcss/typography": "catalog:tailwind3",
+    "@types/lodash-es": "^4.17.12",
     "@types/react": "catalog:",
     "@types/react-dom": "catalog:",
+    "@types/react-is": "^19.2.0",
     "babel-plugin-react-compiler": "catalog:",
     "babel-plugin-styled-components": "^2.3.0",
     "eslint": "^8.57.1",
