+++ packages/visual-editing/src/ui/insert-menu/getSchemaTypeIcon.ts
